requirements.txt
ipdb
ipykernel
jupyter
notebook==6.5.6
numpy==1.23.5
nbdime
pandas==1.5.3
# db-types
google-cloud-storage
pyarrow
pycocotools
opencv-python
scikit-learn==1.3.1
seaborn
keras-cv # in case keras_cv doesnt work, try keras 3
tensorflow-macos==2.15.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.15.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.15.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
tensorflow-datasets==4.9.1 # needed to be able to import keras_cv
pytest
pylint
